infrastructure/app/requirements.txt
boto3==1.35.76
botocore==1.35.76
python-telegram-bot==21.9
langchain==0.3.10
langchain-aws==0.2.9
langchain-community==0.3.9
chromadb==0.5.23
PyPDF2==3.0.1
python-dotenv==1.0.1